+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * **Job-level loggers** - These loggers are added to and removed from the logging sub-system for the duration of the job only. Relative targets default to the job directory and not the `[app.home]/log` directory.
 * **Prometheus Monitoring** - Added an OpenMetrics endpoint `/metrics` that Prometheus can use to collect `Service` operation and health metrics.
+* **RunJob Listener** - This listener will kick off jobs after during the `onMap()` event passing target frame fields in the symbol table of the called jobs. 
 
 ### Fixed
 * **JdbcWriter NPE** - Fixed the issue with the prepared statement not being (re)created when the `Job` is closed and later re-opened as is the case with `Job` instances running in a `Service`.